Potential candidate for release.
Overhaul of the deductions calculations with the addition of RuleFiles to store the lower and upper bounds as well as a maths expresion.

Unit Tests updated to suit updated Salary class constructor. Tests themselves remained the same. Aided the rest of development.

Many comments added.

Tidy up of exception handling.